test/cqlpy: verify LWT reads use the caller's service level
Consolidate the test into the existing test_lwt.py instead of creating a
new file, and address review feedback:

- Separate the retry loop (probing for service level propagation) from
  the actual measurement, so the test structure is clearer.
- Increase N from 20 to 1000 for better signal-to-noise ratio against
  background reads.
- Assert default_delta == 0 (no reads leak to sl:default) instead of
  the weaker default_delta < sl_delta.
- Remove the 0.5s sleep between retries; the probe retries immediately.
- Clarify the SCYLLADB-2379 reference comment to indicate the test
  disproves a theory rather than confirming a bug.
- The Copilot review comments about raise_for_status() and label
  parsing were already addressed in a prior commit.

Refs SCYLLADB-2379.
